systemd/greybox.service: improved start/stop/restart behavior
Add an "ExecStartPre" loop waiting for core-daemon to begin listening
on 127.0.0.1:4038. While this enables a cleaner expression of greybox's
dependency on core-daemon, it would be nice if in the future the latter
would support socket activations (i.e., if we could pass it an already
connected client from within systemd).

Add a loop waiting for all "veth*" and "b.*" interfaces to go away to
the "ExecStop" command. This is hackish, but enables restarting of the
greybox service via "systemctl".

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>