contrib/bitcoin: add sample map and hosts file for bitcoin scenario
This example showcases a completely self-contained bitcoin network,
complete with miner nodes competing against each other.

The example is shipped under 'contrib' since it depends on features
not available with official upstream bitcoin software, such as the
ability of nodes operating in 'regtest' mode to automatically reach
out and connect to their neighbors, as well as reverting the upstream
removal of CPU-based mining.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>