maps/iodine-demo: set `qname-minimization` to `strict`
The current default (`relaxed`) no longer works with iodine
as of bind-9.16.* (bind upstream commit ae52c21), the symptom
being that iodine client queries are first replaced with
minimized substitutes, which fail, and then original queries
are forwarded too late for the tunnel to be successfully negotiated.

This may be related to the way Topgen serves A records directly from
its TLD server views in most cases.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>